Move diagram tools and toolbox (placement) to diagram package
