+++ gaphor/diagram/tests/test_diagramtoolbox.py
@@ -1,9 +1,9 @@
 from gi.repository import Gtk
 
 from gaphor import UML
+from gaphor.diagram.diagramtoolbox import TOOLBOX_ACTIONS
 from gaphor.tests.testcase import TestCase
 from gaphor.ui.diagrampage import DiagramPage
-from gaphor.ui.diagramtoolbox import TOOLBOX_ACTIONS
 
 
 class WindowOwner:
