+++ gaphor/diagram/tests/test_diagramtools.py
